[EuiTableRowCell] Set up Emotion styles + move some "row" styles that belong in here
