+++ src/components/table/_responsive.scss
@@ -45,11 +45,6 @@
       }
     }
 
-    .euiTableRowCell {
-      min-width: 50%;
-      border: none;
-    }
-
     .euiTableRowCellCheckbox {
       border: none;
     }
